Passing the baton with monitors
This is a program that using the "passing the baton" strategy when working in parallel. I'm using monitors , and the rules are as follows:

 - Men and women can't both be in the bathroom, either only men or women can go at a time.
 - There are only 6 stalls in the bathroom, so only 6 people can use at once
 - The program must be as FAIR as possible, meaning everyone must enter and exit the bathroom at some point

This is solving the same bathroom problem as my other program (PG4), but there are 4 other files that must be used with the three here for it to work. The 4 other files are custom built semaphore and monitor classes. These were written by my professor so I'm not posting them hear, I don't have permission to post his work in this repository.